Add RoomMessage type.
Not happy about it, see TODOs.
There are situations where an event that is technically invalid,
still needs to be put through the protections system.
Because event validation sadly is not a part of authorization rules for a server,
so all these events that could be invalid, will still appear and will be used
as an abuse vector.