Fix incorrect schema `PolicyContent`.
All room versions require the PDU to have the content key, and
the value should always be `{}` or an object with properties.

Also to watch out for: there is a nasty TypeBox bug whereby
Composite objects that share the same fields without you
explicitly choosing which one to compose will cause runtime problems.