Remove `TrackedStateEvent`.
Handling redactions of events that have keys preserved in auth rules
made it difficult to continue. We would have had to add key counts to
`TrackedStateEvent` or remove it entirely.

Since removing it entirely was an option, I decided to look into what
the benefits of `TrackedStateEvent` were since the surrounding code
is a little complex and unimplemented (we didn't copy fields from
state events to create minimal tracked events, as was the intention).

https://github.com/Gnuxie/matrix-protection-suite/issues/6

So after looking at the merits it's clear that it's probably not worth
having, because it isn't saving very much and there's a good chance
if it did work as intended, people would make nasty hidden mistakes.