Add handleTimelineEvent to MemberBanSynchronisationProtection.
This needs to be done because state changes will be calculated at a
slower pace than timeline events, and it is often critical to ban
new joiners quickly.

In future it should be a setting if for some reason we get a new
appservice api to push state deltas to clients or we have a client
that can receive state deltas.

We also test that this behaviour is present, though not comprehensively.