Allow for MjolnirEnabledProtectioncConfig to be manipulated.
We needed a way to allow Draupnir to turn on protections by default
on first use, or when upgrading from Mjolnir.
So we added a `enabledProtectionsMigration` member to the config
manager that Draupnir can define.

We removed the `PersistentData` manager since it is now redundant
with the `SchemedMatrixData` manager. Although, it might be
necessary to take a look at it later if we need a data store
for some things e.g. caching room state between startups.

https://github.com/Gnuxie/matrix-protection-suite/issues/7