Introduce concept of preemptive join for ClientRooms.
We need a way to mark a room as preemptively joined.
When clients create new rooms, or use an api call to join them,
they need to immediately take action on the room as soon as
they get a response, not wait for the server to send their
join event down the timeline.
This isn't perfect yet. I'm committing progress here because
I need to decide to change this further.

We could decide that it's not worth distinguishing between
a canonical join and a preemptive join, but a join in
ClientRooms has a specific meaning, that they are canonically
joined to the room via a presence in /joined_rooms.

A preemptive join violates that expectation, and a client shouldn't
be given timeline events for a room that is only preemptively joined.

Also, preemptive joins are going to be a source of introduced error
so it's important to be able to warn about when they are being
used inappropriately (for example, we can detect when someone
has used it but there isn't a corresponding entry in the next
call to `/joined_rooms`.