Move preemptive join management to ClientRooms.
This is in order to keeep the joined rooms revision strictly tied
to what the homeserver most recently responded to `/joined_rooms`
with.

We now need to remove `addClientRooms` from the ClientsInRoomMap
and make it construct ClientRooms.

We now have an issue that a ClientRoom can preemptively join
a room before it is added to the ClientsInRoomMap, meaning
that we can't index the rooms properly.

We'd do this instead of expsoing the list of preemptively joined
rooms on `ClientRooms` because it's possible the RoomStateManager
is calling ClientsInRoomMap['getManagedClients'] before calling
`addClientRooms` but i'm less sure of this honestly.