Add StandardPolicyRoomEditor.
This is derived from the BotSDKPolicyRoomEditor from MPSForBotSDK.
We managed this by using the client capability interface for sending
state events and also using the roomStateRevisionIssuer to check
for policies to delete. We use the state revision to do this and
not the policy revision because we haven't really figured out
what to do with mjolnir's old normalisation of policies behaviour.

MPS normalises policies in revisions, and i'm not sure that's correct.
Either way all versions need to be removed to unban somebody.