ClientsInRoomMap should be constructing the ClientRoom's interned.
We also removed a weird out of place optimisation to preload state of
joined rooms, since this can be moved to client specific code.
And it means we don't have a dependency on the room state manager
which would have caused a bootstrapping problem later if we need to
merge more concepts.