Rework ConsequenceProvider to be a set of Capability Providers.
This is a big change but was necessary for where we want to go.
We brought this work forwards because we needed to have consistent
renderers for consequences, which was going to be difficult
to do without a good conceptualisation of consequences as
capabilities.

We haven't written the renderers Draupnir side yet,
but when we do, we will need to verify that they are
mixed into the capabilities dynamically, not eagerly.
Ie, when the capability provider factories are called.