PowerLevelsMirror
Need to change all caps and protections to give requiredEvent and requiredState permissions
probably as an object that labels them explicitly. So a breaking change.

Then later we need to change the user consequences protection to take a revision
and apply it to a room. Probably renaming the stuff that accepts a revision.
Though this should be done in a follow up PR.