Minor change to consistency of MissinPL callback.
I don't like this because it feels like the CB will get spammed
when lots of rooms and lots of protections don't have
permissions.

Of course, there's a PL change hook so it can probably be
changed to incorporate every protection change
as a list.

Then there needs to be a hook when someone adds a room
that runs the same check and displays it.