Consistency with SetState/SetMembership.
We add a `ProtectedRoomsManager` which splits the role of the
`ProtectedRoomsConfig` so that the `ProtectedRoomsConfig` is soley
focussed on loading and parsing account data. Whereas the manager
is focussed on taking the parsed account data and consistently
intialising and adding/removing room from the set.

Fixes https://github.com/Gnuxie/matrix-protection-suite/issues/34.