Request RoomStateRevisionIssuers in order for protected rooms.
We believe this is the cause of a performance regression in Draupnir:
https://github.com/the-draupnir-project/Draupnir/issues/380.
Unfortunately we don't have the infrastructure in our unit testing
or deployment to be able to truthfully understand the cause.

So we have created a follow up https://github.com/the-draupnir-project/Draupnir/issues/389.
Our approach for now is just to guess some of the most glaring
changes that have caused the regression.

If this fails to make a significant difference, other culprits
are changes to the parsing code (ie, there was a recent version
of MPS where we moved away from accidentally using the empty
schema for most event content), including the recent upgrade
to typebox 0.32.22.