UNSTAGE whatever - I've decided i'm had enough.
Config description needs to be its own thing and it needs
to be based on Schema and JSONPath.
We need to build this oriented around a unit test
that we can be sure we're getting accurate path information from.

The set/add remove commands need to work dynamically with any
property and then validate after the fact.

We can get set semantics by having uniqueItems: true
we need to test for that and see how it breaks
alternatively a transform can be used to create sets.

The description needs to be able to parse a config
and add a mirror hatch to it that we can use to access values
through the ype safe interface or json interface.