Split ProtectionsManager into 3 responsibilities and fake inline.
We want to be able to fake the entire protections manager without
using a different class just by providing mock backends to all
of the dependancies that do IO.

We're now stuck and hesitating because we need to change the way
we run intialization code in static methods ie the `create` methods.

So yeah I want to do something about that, hopefully in a way that
is not shit.