Add Cleaner layer to replace PolicyListConfig in ProtectedRoomsSet.
The `PolicyListConfig` is doing the job of organising revisions, and
revision issuers alongside persisting the config. This was bad because
the process of creating the revision issuers and the aggregated
revision issuer is going to be quite complicated with agree/disagree.

Additionally there is a utility in Draupnir called
`getPolicyListsInfo` which tried to match watched lists up with
revisions.
When the new component `WatchedPolicyRooms` will be doing that job.

https://github.com/the-draupnir-project/planning/issues/4