This would work but it's taking too much time to work out.
There are a few ideas in this branch to support protections.
1. Is notifying protections for each room where a member
matches policies.. and also giving that revision to hand
for protection startup.

2. was something about allowing you to find the rules matching
members who have been in the community but aren't any more.
That kind of needs to be a distinct revision and i don't
think it needs to include the rooms all the time.