Rework pagination.
It's important to make pagination cleaner and isolate state from other
concerns. State has been encapsulated into PaginationSession and
PaginationIterator.

We have deleted `RoomEventRelationsGetter` for now, because we don't
have a use for it and it's not a priority to move to the new system.

Instead we have provided `/messages` pagination instead, which we have
an immediate use for in improving redaction in draupnir.