Rework event mixin extraction.
Now we use a description object to provide the parser and shape for
each mixin that we want to parse. This is great because it allows you
to use the description object as the type tag for the mixin, and
therefore the shape information will be available with that.  This
allows mixins to be defined in a modular way without aggregating all
known mixins into one enum type tag or one union type with all shapes.
I'm pretty happy with it!!!