Implement MemberBanIntentProjection while tuning Projection types.
I spent way too long figuring out how to make this thing work, only to
realise the input extractor was expanding to never[] instead of never
or unknown[] instead of unknown or vice versa. Quite a simple bug in
the end.

We also make the data accessor parts of projection nodes a mixin on
the projection node type definition. This was to keep the type checker
happy when returning subtypes from the reduceDelta method.

https://github.com/the-draupnir-project/planning/issues/2