Add intent projection to member ban sync.
Yesterday we tried to make the protection actually run directly from
the projection, but this quickly turned out to require too deep of a
refactor. Specifically, we would have to bridge more revisions to the
intent projection, and also make up for a lack of information in
others, such as room membership.
https://github.com/the-draupnir-project/planning/issues/2